requirements.txt
amqp==5.0.1
beautifulsoup4==4.9.2
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.7
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5

psycopg2~=2.8.6
pip~=20.1.1
configparser~=5.0.0
wheel~=0.34.2
setuptools~=49.2.0